+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # ChangeLog
 
-* fix transform
+* add compute shader support
+* add select(ternary operation) node
+* add null node